+++ src/statusbar/StatusBar.ts
@@ -1,9 +1,10 @@
 //	Imports ____________________________________________________________________
 
-import { basename } from 'path';
 import * as vscode from 'vscode';
 
+import { formatLabel } from '../@l13/formats';
 import { isMacOs, isWindows } from '../@l13/platforms';
+
 import { Project } from '../@types/workspaces';
 
 import * as settings from '../common/settings';
@@ -46,7 +47,7 @@ export class StatusBar {
 			const icon = settings.isCodeWorkspace(workspacePath) ? 'submodule' : 'directory';
 			const name:string = this.getProjectName('projects', workspacePath) || this.getProjectName('favorites', workspacePath);
 			
-			this.statusBarItem.text = `$(file-${icon}) ${name || basename(workspacePath, '.code-workspace')}`;
+			this.statusBarItem.text = `$(file-${icon}) ${name || formatLabel(workspacePath)}`;
 		}
 		
 	}